backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
reportlab
python-multipart